GoEuro technical test
This is a simple project requested as part of the application process to
work for GoEuro.  It consists of a command line application which makes
a REST request on the GoEuro API and gets a list of JSON objects as a
result.  These are then written to a CSV file.